requirements.txt
fastapi
pydantic
Pillow
numpy
torch
torchvision
matplotlib
scipy
torchcam
uvicorn
